requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==4.2.15
django-filter==24.3
django-widget-tweaks==1.5.0
filelock==3.15.4
gunicorn==23.0.0
packaging==24.1
pbr==6.0.0
pillow==10.4.0
platformdirs==4.2.2
South==1.0.2
sqlparse==0.5.1
stevedore==5.2.0
typing_extensions==4.12.2
tzdata==2024.1
virtualenv==20.26.3
virtualenv-clone==0.5.7
virtualenvwrapper==6.1.0
